requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-modeltranslation==0.18.10
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-smtp-ssl==1.0
idna==3.4
markdown2==2.4.8
Pillow==9.5.0
psycopg2-binary==2.9.6
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
